Responsive desogn completed
