+++ app/admins/page-client.tsx
@@ -75,7 +75,7 @@ export default function AdminsPageClient() {
   return (
     <div className="flex min-h-screen flex-col">
       <Sidebar />
-      <div className="flex-1 md:ml-[var(--sidebar-width)] lg:-mt-12 -mt-8">
+      <div className="flex-1 md:ml-[calc(var(--sidebar-width)-40px)] md:-mt-12 -mt-8">
         <main className="flex-1 space-y-4 p-4 md:p-8 pt-6">
           <div className="flex items-center justify-between">
             <h1 className="text-xl md:text-3xl font-bold tracking-tight">Admin Management</h1>
