+++ app/auctions/page-client.tsx
@@ -13,12 +13,12 @@ export default function AuctionsPageClient() {
   return (
     <div className="flex min-h-screen flex-col">
       <Sidebar />
-      <div className="flex-1 md:ml-[var(--sidebar-width)] lg:-mt-12 -mt-8">
+      <div className="flex-1 md:ml-[calc(var(--sidebar-width)-40px)] md:-mt-12 -mt-8">
         <main className="flex-1 space-y-4 p-4 md:p-8 pt-6">
           <div className="flex items-center justify-between">
             <h1 className="text-xl md:text-3xl font-bold tracking-tight">Auction Management</h1>
             <div className="flex items-center gap-2">
-              <span className="text-sm text-muted-foreground">Total Auctions: 342</span>
+              <span className="text-sm text-muted-foreground"><span className="hidden md:block">Total</span>Auctions: 342</span>
             </div>
           </div>
