+++ app/dashboard/page.tsx
@@ -419,9 +419,9 @@ export default function DashboardPage() {
   )
 
   return (
-    <div className="flex min-h-screen flex-col overflow-x-hidden">
+    <div className="flex min-h-screen flex-col">
       <Sidebar />
-      <div className="flex-1 md:ml-[var(--sidebar-width)] -mt-8 md:-mt-8">
+      <div className="flex-1 md:ml-[calc(var(--sidebar-width)-40px)] md:-mt-12 -mt-8">
         <main className="flex-1 space-y-4 p-4 md:p-8 pt-6">
           <div className="flex items-center justify-between">
             <h1 className="text-xl md:text-3xl font-bold tracking-tight">Dashboard</h1>
