+++ app/products/page-client.tsx
@@ -151,19 +151,28 @@ export default function ProductsPageClient() {
   return (
     <div className="flex min-h-screen flex-col">
       <Sidebar />
-      <div className="flex-1 md:ml-[var(--sidebar-width)] lg:-mt-12 -mt-8">
+      <div className="flex-1 md:ml-[calc(var(--sidebar-width)-40px)] md:-mt-12 -mt-8">
         <main className="flex-1 space-y-4 p-4 md:p-8 pt-6">
-          <div className="flex items-center justify-between">
-            <h1 className="text-xl md:text-3xl font-bold tracking-tight">Product Management</h1>
-            <div className="flex items-center gap-2">
-              <span className="text-sm text-muted-foreground hidden md:block">Total Products: {totalProducts}</span>
-              <Button variant="outline" size="sm" onClick={() => setShowFilters(!showFilters)}>
-                <Filter className="mr-2 h-4 w-4" />
+          <div className="flex flex-wrap items-center justify-between gap-4 md:gap-6">
+            <h1 className="text-2xl md:text-4xl font-semibold tracking-tight text-gray-900 dark:text-white">
+              Product Management
+            </h1>
+            <div className="flex items-center gap-4">
+              <span className="text-sm md:text-base text-muted-foreground">
+                <span className="hidden md:inline">Total </span>Products: 
+                <span className="font-medium text-gray-800 dark:text-gray-200"> {totalProducts}</span>
+              </span>
+              <Button
+                variant="outline"
+                size="sm"
+                className="flex items-center gap-2 px-3 py-1.5 text-sm md:text-base"
+                onClick={() => setShowFilters(!showFilters)}
+              >
+                <Filter className="h-4 w-4 md:h-5 md:w-5" />
                 {showFilters ? "Hide Filters" : "Show Filters"}
               </Button>
             </div>
           </div>
-
           <div className="flex gap-4 flex-row md:items-center md:justify-between">
             <div className="flex flex-1 items-center gap-2">
               <div className="relative flex-1 md:max-w-sm">
@@ -199,8 +208,8 @@ export default function ProductsPageClient() {
             />
           )}
 
-          <Card>
-          <CardHeader className="p-4 text-sm md:text-base lg:text-lg">
+          <Card className="-ml-5 md:ml-0 -mr-4 md:mr-0 overflow-x-hidden">
+            <CardHeader className="p-4 text-sm md:text-base lg:text-lg">
              <CardTitle>
                 {selectedTab === "active" 
                   ? "All Products" 
@@ -217,76 +226,78 @@ export default function ProductsPageClient() {
               </CardDescription>
             </CardHeader>
             <CardContent className="p-0">
-              <Table>
-              <TableHeader>
-                <TableRow>
-                  <TableHead>Product Name</TableHead>
-                  <TableHead className="hidden md:table-cell">Category</TableHead>
-                  <TableHead>Price</TableHead>
-                  <TableHead className="hidden md:table-cell">Merchant</TableHead>
-                  <TableHead>Status</TableHead>
-                  <TableHead className="hidden md:table-cell">Stock</TableHead>
-                  <TableHead className="hidden md:table-cell">Rating</TableHead>
-                </TableRow>
-              </TableHeader>
-              <TableBody>
-                {isLoadingData ? (
-                  <TableRow>
-                    <TableCell colSpan={7} className="h-24 text-center">
-                      Loading products...
-                    </TableCell>
-                  </TableRow>
-                ) : products.length === 0 ? (
-                  <TableRow>
-                    <TableCell colSpan={7} className="h-24 text-center">
-                      No products found.
-                    </TableCell>
-                  </TableRow>
-                ) : (
-                  products.map((product) => (
-                    <TableRow
-                      key={product._id}
-                      className="cursor-pointer hover:bg-muted/50"
-                      onClick={() => setSelectedProduct(product)}
-                    >
-                      <TableCell className="font-medium">{product.productName}</TableCell>
-                      <TableCell className="hidden md:table-cell">{product.category.categoryName}</TableCell>
-                      <TableCell>${product.price.toFixed(2)}</TableCell>
-                      <TableCell className="hidden md:table-cell">{product.merchantDetail.merchantName}</TableCell>
-                      <TableCell>
-                        {selectedTab === "deleted" ? (
-                          <span className="flex items-center text-gray-500">
-                            <Trash2 className="mr-1 h-4 w-4" /> Deleted
-                          </span>
-                        ) : product.isBanned ? (
-                          <span className="flex items-center text-red-500">
-                            <XCircle className="mr-1 h-4 w-4" /> Banned
-                          </span>
-                        ) : (
-                          <span className="flex items-center text-green-500">
-                            <CheckCircle className="mr-1 h-4 w-4" /> Active
-                          </span>
-                        )}
-                      </TableCell>
-                      <TableCell className="hidden md:table-cell">
-                        {product.quantity} ({product.soldQuantity} sold)
-                      </TableCell>
-                      <TableCell className="hidden md:table-cell">
-                        {product.avgRating > 0 ? (
-                          <div className="flex items-center">
-                            {product.avgRating}
-                            <span className="text-yellow-500 ml-1">★</span>
-                            <span className="text-xs text-muted-foreground ml-1">({product.review.length})</span>
-                          </div>
-                        ) : (
-                          <span className="text-muted-foreground">No ratings</span>
-                        )}
-                      </TableCell>
+              <div className="overflow-x-hidden">
+                <Table>
+                  <TableHeader>
+                    <TableRow>
+                      <TableHead>Product Name</TableHead>
+                      <TableHead className="hidden md:table-cell">Category</TableHead>
+                      <TableHead>Price</TableHead>
+                      <TableHead className="hidden md:table-cell">Merchant</TableHead>
+                      <TableHead>Status</TableHead>
+                      <TableHead className="hidden md:table-cell">Stock</TableHead>
+                      <TableHead>Rating</TableHead>
                     </TableRow>
-                  ))
-                )}
-              </TableBody>
-              </Table>
+                  </TableHeader>
+                  <TableBody>
+                    {isLoadingData ? (
+                      <TableRow>
+                        <TableCell colSpan={7} className="h-24 text-center">
+                          Loading products...
+                        </TableCell>
+                      </TableRow>
+                    ) : products.length === 0 ? (
+                      <TableRow>
+                        <TableCell colSpan={7} className="h-24 text-center">
+                          No products found.
+                        </TableCell>
+                      </TableRow>
+                    ) : (
+                      products.map((product) => (
+                        <TableRow
+                          key={product._id}
+                          className="cursor-pointer hover:bg-muted/50"
+                          onClick={() => setSelectedProduct(product)}
+                        >
+                          <TableCell className="font-medium">{product.productName}</TableCell>
+                          <TableCell className="hidden md:table-cell">{product.category.categoryName}</TableCell>
+                          <TableCell>${product.price.toFixed(2)}</TableCell>
+                          <TableCell className="hidden md:table-cell">{product.merchantDetail.merchantName}</TableCell>
+                          <TableCell>
+                            {selectedTab === "deleted" ? (
+                              <span className="flex items-center text-gray-500">
+                                <Trash2 className="mr-1 h-4 w-4" /> Deleted
+                              </span>
+                            ) : product.isBanned ? (
+                              <span className="flex items-center text-red-500">
+                                <XCircle className="mr-1 h-4 w-4" /> Banned
+                              </span>
+                            ) : (
+                              <span className="flex items-center text-green-500">
+                                <CheckCircle className="mr-1 h-4 w-4" /> Active
+                              </span>
+                            )}
+                          </TableCell>
+                          <TableCell className="hidden md:table-cell">
+                            {product.quantity} ({product.soldQuantity} sold)
+                          </TableCell>
+                          <TableCell>
+                            {product.avgRating > 0 ? (
+                              <div className="flex items-center">
+                                {product.avgRating}
+                                <span className="text-yellow-500 ml-1">★</span>
+                                <span className="text-xs text-muted-foreground ml-1">({product.review.length})</span>
+                              </div>
+                            ) : (
+                              <span className="text-muted-foreground">No ratings</span>
+                            )}
+                          </TableCell>
+                        </TableRow>
+                      ))
+                    )}
+                  </TableBody>
+                </Table>
+              </div>
             </CardContent>
           </Card>
