+++ app/signin/page.tsx
@@ -163,7 +163,7 @@ export default function SignInPage() {
 
   const handleResend = async () => {
     try {
-      const res = await fetch("/api/resend-otp", {
+      const res = await fetch("/api/auth/resend-otp", {
         method: "POST",
         headers: { "Content-Type": "application/json" },
         body: JSON.stringify({ email }),
