+++ components/footer.tsx
@@ -2,7 +2,7 @@ export function Footer() {
     const currentYear = new Date().getFullYear()
   
     return (
-      <footer className="mt-auto border-t py-6 ml-0 md:ml-[var(--sidebar-width)]">
+      <footer className="mt-auto border-t py-6 ml-0 md:ml-[calc(var(--sidebar-width)-20px)]">
         <div className="container mx-auto px-4">
           <div className="flex flex-col items-center justify-between gap-4 md:flex-row">
             <div className="text-sm text-muted-foreground">© {currentYear} Marketplace Admin. All rights reserved.</div>
